requirements.txt
openai==1.59.9
pyyaml==6.0.2
python-dotenv==1.0.0
